Fixed errors in Controller. Connect with PostgreSQL

+++ courseworkSpring/src/main/java/main/repository/BalanceRepository.java
@@ -3,6 +3,6 @@ package main.repository;
 import main.entity.Balance;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BalanceRepository extends CrudRepository<Balance, Long>
+public interface BalanceRepository extends CrudRepository<Balance, Integer>
 {
 }
