+++ courseworkSpring/src/main/java/main/repository/OperationRepository.java
@@ -3,6 +3,6 @@ package main.repository;
 import main.entity.Operation;
 import org.springframework.data.repository.CrudRepository;
 
-public interface OperationRepository extends CrudRepository<Operation, Long>
+public interface OperationRepository extends CrudRepository<Operation, Integer>
 {
 }
